refactor: unify geocoding util and move API key to env
