+++ Backend/.env.example
@@ -34,3 +34,5 @@ RAZORPAY_KEY_SECRET=your_razorpay_key_secret
 # Redis (Required)
 # --------------------------------
 REDIS_URI=redis://localhost:6379
+# OpenCage API key for reverse geocoding
+OPENCAGE_API_KEY=your_opencage_api_key_here
